+++ Shared/GR_Code_Events.lua
@@ -19,13 +19,13 @@ function events:StartBaseEvents()
         local _, msg = ...
 
         if not msg then return
-        elseif msg:find(L["PLAYER_JOINED_GUILD"]) then observer:Notify('PLAYER_JOINED_GUILD', ...)
-        elseif msg:find(L["PLAYER_DECLINED_INVITE"]) then observer:Notify('PLAYER_DECLINED_INVITE', msg)
-        elseif msg:find(L["PLAYER_NOT_ONLINE"]) then observer:Notify('PLAYER_NOT_ONLINE', ...)
-        elseif msg:find(L["PLAYER_NOT_PLAYING"]) then observer:Notify('PLAYER_NOT_PLAYING', ...)
-        elseif msg:find(L["PLAYER_NOT_FOUND"]) then observer:Notify('PLAYER_NOT_FOUND', ...)
-        elseif msg:find(L["PLAYER_IN_GUILD"]) then observer:Notify('PLAYER_IN_GUILD', ...)
-        elseif msg:find(L["PLAYER_ALREADY_IN_GUILD"]) then observer:Notify('PLAYER_ALREADY_IN_GUILD', ...)
+        elseif msg:match(L["PLAYER_JOINED_GUILD"]) then observer:Notify('PLAYER_JOINED_GUILD', ...)
+        elseif msg:match(L["PLAYER_DECLINED_INVITE"]) then observer:Notify('PLAYER_DECLINED_INVITE', msg)
+        elseif msg:match(L["PLAYER_NOT_ONLINE"]) then observer:Notify('PLAYER_NOT_ONLINE', ...)
+        elseif msg:match(L["PLAYER_NOT_PLAYING"]) then observer:Notify('PLAYER_NOT_PLAYING', ...)
+        elseif msg:match(L["PLAYER_NOT_FOUND"]) then observer:Notify('PLAYER_NOT_FOUND', ...)
+        elseif msg:match(L["PLAYER_IN_GUILD"]) then observer:Notify('PLAYER_IN_GUILD', ...)
+        elseif msg:match(L["PLAYER_ALREADY_IN_GUILD"]) then observer:Notify('PLAYER_ALREADY_IN_GUILD', ...)
         end
     end